+++ CHANGELOG.md
@@ -1,4 +1,14 @@
 # [Hyperledger Burrow](https://github.com/hyperledger/burrow) Changelog
+## [0.23.1] - 2018-11-14
+### Fixed
+- [EVM] state/Cache no longer allows SetStorage on accounts that do not exist
+- [GRPC] GetAccount on unknown account no longer causes a panic
+
+### Added
+- [Docker] Added solc 0.4.25 binary to docker container so that burrow deploy has what it needs to function
+- [Execution] panics from executors are captured and pushed to error sink of TxExecution
+
+
 ## [0.23.0] - 2018-11-09
 ### Changed
 - [ABI] provides fast event lookup of EventID
@@ -311,6 +321,7 @@ This release marks the start of Eris-DB as the full permissioned blockchain node
   - [Blockchain] Fix getBlocks to respect block height cap.
 
 
+[0.23.1]: https://github.com/hyperledger/burrow/compare/v0.23.0...v0.23.1
 [0.23.0]: https://github.com/hyperledger/burrow/compare/v0.22.0...v0.23.0
 [0.22.0]: https://github.com/hyperledger/burrow/compare/v0.21.0...v0.22.0
 [0.21.0]: https://github.com/hyperledger/burrow/compare/v0.20.1...v0.21.0